Question_02
Sorting of 2D array in both the manner i.e. in ascending and descending order, by the means of switch case operation an option which order sort should be provided to user. Instructions to follow 1. All the elements of array should be entered by user no pre filled values are acceptable. 2. Use of Inbuilt functions are restricted.